Refactoring of instrument classes and handling of dynamic methods from YAML
We still need to define all the abstact methods in python. The use of wrappers comes handy to help the user know when a base method is defined inm the YAML specification, and to allow for them to share the same name. However it would still better to allow for a driver to be fully specified in the YAML file (including abstratc methods) so that no python code is needed.